pyproject.toml
[project]
name = "image-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flet>=0.80.1",
    "pillow>=12.1.0",
]
